+++ sqleto/lib/sqleto.dart
@@ -5,5 +5,6 @@ export 'src/sqleto_config.dart';
 export 'src/where.dart';
 export 'src/exceptions/sqleto_exception.dart';
 export 'src/enums/operator.dart';
+export 'src/utils/sqleto_utils.dart';
 
 export 'package:sqleto_annotation/sqleto_annotation.dart';