+++ sqleto/lib/src/utils/sqleto_schema_utils.dart
@@ -5,49 +5,51 @@ import 'package:sqleto/sqleto.dart';
 class SQLetoSchemaUtils {
   SQLetoSchemaUtils._();
 
-  static bool isValidSchema(Type T) {
-    final cm = reflectClass(T);
-    return cm.isSubtypeOf(reflectClass(SQLetoSchema));
+  static bool isValidSchema(Object Function() object) {
+    final im = reflect(object());
+    return im.type.isSubtypeOf(reflectClass(SQLetoSchema));
   }
 
-  static String tableName(Type T) {
-    return invokeFromMap(T, {}).tableName;
+  static String tableName(Object Function() object) {
+    if (!isValidSchema(object)) throw InvalidSchemaException('Its given class does not extends from Schema!');
+
+    return invokeFromMap(object, {}).tableName;
   }
 
-  static String buildSELECT(Type T, [String? whereScript]) {
-    if (!isValidSchema(T)) throw InvalidSchemaException('Its given class does not extends from Schema!');
+  static String buildSELECT(Object Function() object, [String? whereScript]) {
+    if (!isValidSchema(object)) throw InvalidSchemaException('Its given class does not extends from Schema!');
 
-    return 'SELECT * FROM ${tableName(T)}${whereScript ?? ''}';
+    return 'SELECT * FROM ${tableName(object)}${whereScript ?? ''}';
   }
 
-  static String getPKName(Type T) {
-    if (!isValidSchema(T)) throw InvalidSchemaException('Its given class does not extends from Schema!');
+  static String getPKName(Object Function() object) {
+    if (!isValidSchema(object)) throw InvalidSchemaException('Its given class does not extends from Schema!');
 
-    final cm = reflectClass(T);
+    final im = reflect(object());
 
-    final base = cm.superclass!.typeArguments.first as ClassMirror;
+    final base = im.type.superclass as ClassMirror;
 
     final vms = base.declarations.values.whereType<VariableMirror>().toList();
 
     if (vms.any((e) => _instanceMirrorWithFieldAnnotation(e).getField(#primaryKey).reflectee == true)) {
       return vms.firstWhere((e) => _instanceMirrorWithFieldAnnotation(e).getField(#primaryKey).reflectee == true).simpleName.name;
     } else {
-      throw InvalidSchemaException('T schema $T does not have a primary key!');
+      throw InvalidSchemaException('T schema ${object().runtimeType} does not have a primary key!');
     }
   }
 
-  static String buildDELETE(Type T) {
-    if (!isValidSchema(T)) throw InvalidSchemaException('Its given class does not extends from Schema!');
+  static String buildDELETE(Object Function() object) {
+    if (!isValidSchema(object)) throw InvalidSchemaException('Its given class does not extends from Schema!');
 
-    return 'DELETE FROM ${tableName(T).toUpperCase()} WHERE ${getPKName(T).toUpperCase()} = @${getPKName(T)} RETURNING *';
+    return 'DELETE FROM ${tableName(object).toUpperCase()} WHERE ${getPKName(object).toUpperCase()} = @${getPKName(object)} RETURNING *';
   }
 
-  static String buildUPDATE(Type T) {
-    if (!isValidSchema(T)) throw InvalidSchemaException('Its given class does not extends from Schema!');
+  static String buildUPDATE(Object Function() object) {
+    if (!isValidSchema(object)) throw InvalidSchemaException('Its given class does not extends from Schema!');
 
-    final cm = reflectClass(T);
+    final im = reflect(object());
 
-    final base = cm.superclass!.typeArguments.first as ClassMirror;
+    final base = im.type.superclass as ClassMirror;
 
     final vms = base.declarations.values.whereType<VariableMirror>();
 
@@ -55,7 +57,7 @@ class SQLetoSchemaUtils {
 
     final setters = <String>[];
 
-    builder.add('UPDATE ${tableName(T).toUpperCase()} SET');
+    builder.add('UPDATE ${tableName(object).toUpperCase()} SET');
 
     for (var vm in vms) {
       final im = _instanceMirrorWithFieldAnnotation(vm);
@@ -67,25 +69,23 @@ class SQLetoSchemaUtils {
 
     builder.add(setters.join(', '));
 
-    final pkName = getPKName(T);
+    final pkName = getPKName(object);
 
     builder.add('WHERE $pkName = @$pkName RETURNING *');
 
     return builder.join(' ');
   }
 
-  static String buildINSERT(SQLetoSchema Function() schema) {
-    final type = schema().runtimeType;
-
-    if (!isValidSchema(type)) throw InvalidSchemaException('Its given class does not extends from Schema!');
+  static String buildINSERT(Object Function() object) {
+    if (!isValidSchema(object)) throw InvalidSchemaException('Its given class does not extends from Schema!');
 
     final builder = <String>[];
 
     builder.add('INSERT INTO');
 
-    builder.add(tableName(type).toUpperCase());
+    builder.add(tableName(object).toUpperCase());
 
-    final im = reflect(schema());
+    final im = reflect(object());
 
     final cm = im.type;
 
@@ -106,12 +106,12 @@ class SQLetoSchemaUtils {
             {
               final value = im.getField(vm.simpleName).reflectee;
 
-              if (value is! String) throw InvalidSchemaException('$type field >>${vm.simpleName.name.toUpperCase()}<< is marked as email but is not a String/TEXT!');
+              if (value is! String) throw InvalidSchemaException('${object().runtimeType} field >>${vm.simpleName.name.toUpperCase()}<< is marked as email but is not a String/TEXT!');
 
               final regex = RegExp(r"^[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?$");
 
               if (!regex.hasMatch(value)) {
-                throw InvalidSchemaException('$type field >>${vm.simpleName.name.toUpperCase()}<< is marked as email but the value given is not an email!');
+                throw InvalidSchemaException('${object().runtimeType} field >>${vm.simpleName.name.toUpperCase()}<< is marked as email but the value given is not an email!');
               }
 
               break;
@@ -120,12 +120,13 @@ class SQLetoSchemaUtils {
             {
               final value = im.getField(vm.simpleName).reflectee;
 
-              if (value is! String) throw InvalidSchemaException('$type field >>${vm.simpleName.name.toUpperCase()}<< is marked as username but is not a String/TEXT!');
+              if (value is! String) throw InvalidSchemaException('${object().runtimeType} field >>${vm.simpleName.name.toUpperCase()}<< is marked as username but is not a String/TEXT!');
 
               final regex = RegExp(r"^[a-zA-Z0-9_]{5,24}$");
 
               if (!regex.hasMatch(value)) {
-                throw InvalidSchemaException('$type field >>${vm.simpleName.name.toUpperCase()}<< is marked as username, only letters, numbers and _ are allowed, max 24 characters!');
+                throw InvalidSchemaException(
+                    '${object().runtimeType} field >>${vm.simpleName.name.toUpperCase()}<< is marked as username, only letters, numbers and _ are allowed, max 24 characters!');
               }
               break;
             }
@@ -133,10 +134,10 @@ class SQLetoSchemaUtils {
             {
               final value = im.getField(vm.simpleName).reflectee;
 
-              if (value is! String) throw InvalidSchemaException('$type field >>${vm.simpleName.name.toUpperCase()}<< is not a String/TEXT!');
+              if (value is! String) throw InvalidSchemaException('${object().runtimeType} field >>${vm.simpleName.name.toUpperCase()}<< is not a String/TEXT!');
 
               if (value.isEmpty) {
-                throw InvalidSchemaException('$type field >>${vm.simpleName.name.toUpperCase()}<< cannot be empty!');
+                throw InvalidSchemaException('${object().runtimeType} field >>${vm.simpleName.name.toUpperCase()}<< cannot be empty!');
               }
               break;
             }
@@ -144,10 +145,10 @@ class SQLetoSchemaUtils {
             {
               final value = im.getField(vm.simpleName).reflectee;
 
-              if (value is! num) throw InvalidSchemaException('$type field >>${vm.simpleName.name.toUpperCase()}<< is marked as not negative number but is not a number!');
+              if (value is! num) throw InvalidSchemaException('${object().runtimeType} field >>${vm.simpleName.name.toUpperCase()}<< is marked as not negative number but is not a number!');
 
               if (value.isNegative) {
-                throw InvalidSchemaException('$type field >>${vm.simpleName.name.toUpperCase()}<< cannot be negative!');
+                throw InvalidSchemaException('${object().runtimeType} field >>${vm.simpleName.name.toUpperCase()}<< cannot be negative!');
               }
               break;
             }
@@ -155,12 +156,12 @@ class SQLetoSchemaUtils {
             {
               final value = im.getField(vm.simpleName).reflectee;
 
-              if (value is! String) throw InvalidSchemaException('$type field >>${vm.simpleName.name.toUpperCase()}<< is not a String/UUID!');
+              if (value is! String) throw InvalidSchemaException('${object().runtimeType} field >>${vm.simpleName.name.toUpperCase()}<< is not a String/UUID!');
 
               final regex = RegExp(r"^[0-9A-Za-z]{8}-[0-9A-Za-z]{4}-4[0-9A-Za-z]{3}-[89ABab][0-9A-Za-z]{3}-[0-9A-Za-z]{12}$");
 
               if (!regex.hasMatch(value)) {
-                throw InvalidSchemaException('$type field >>${vm.simpleName.name.toUpperCase()}<< is not a valid UUID!');
+                throw InvalidSchemaException('${object().runtimeType} field >>${vm.simpleName.name.toUpperCase()}<< is not a valid UUID!');
               }
 
               break;
@@ -180,24 +181,24 @@ class SQLetoSchemaUtils {
 
   static String _buildReferenceFieldName(VariableMirror vm) {
     if (_instanceMirrorWithFieldAnnotation(vm).getField(#password).reflectee == true) {
-      return "crypt(@${_snakeCaseNORMALIZER(vm.simpleName.name)}, gen_salt('bf', 4))";
+      return "encode(digest(@${_snakeCaseNORMALIZER(vm.simpleName.name)}, 'sha256'), 'hex')";
     } else {
       return '@${_snakeCaseNORMALIZER(vm.simpleName.name)}';
     }
   }
 
-  static String buildClassSCHEMA(Type T) {
-    if (!isValidSchema(T)) throw InvalidSchemaException('Its given class does not extends from Schema!');
+  static String buildClassSCHEMA(Object Function() object) {
+    if (!isValidSchema(object)) throw InvalidSchemaException('Its given class does not extends from Schema!');
 
-    final cm = reflectClass(T);
+    final im = reflect(object);
 
-    final base = cm.superclass as ClassMirror;
+    final base = im.type.superclass as ClassMirror;
 
     final vms = base.declarations.values.whereType<VariableMirror>().toList();
 
     final buffer = StringBuffer();
 
-    buffer.writeln('CREATE TABLE IF NOT EXISTS ${tableName(T).toUpperCase()} (');
+    buffer.writeln('CREATE TABLE IF NOT EXISTS ${tableName(object).toUpperCase()} (');
 
     for (var i = 0; i < vms.length; i++) {
       final vm = vms[i];
@@ -257,7 +258,7 @@ class SQLetoSchemaUtils {
     if (references != null) {
       final cm = reflectClass(references);
 
-      final referenceTableName = 'tb_${_snakeCaseNORMALIZER(cm.simpleName.name)}'.toUpperCase();
+      final referenceTableName = 'TB_${_snakeCaseNORMALIZER(cm.simpleName.name).toUpperCase()}';
 
       final vms = cm.declarations.values.whereType<VariableMirror>().toList();
 
@@ -265,8 +266,6 @@ class SQLetoSchemaUtils {
         final referencePkVm = vms.firstWhere((e) => e.metadata.any((e) => e.getField(#primaryKey).reflectee));
 
         final referencePkType = (referencePkVm.metadata.first.getField(#type).reflectee as SQLetoType).name;
-
-        if (referencePkType != type) throw Exception('O PRIMARY KEY DA CLASSE REFERENCIADA NÃO É DO MESMO TIPO DA CAMPO DA CLASSE CRIADA');
 
         if (referencePkType != type) throw InvalidReferencedSchemaException('The referenced schema primary key type is not the same to the given Schema!');
 
@@ -296,28 +295,30 @@ class SQLetoSchemaUtils {
     }
   }
 
-  static dynamic invokeFromMap(Type T, Map<String, dynamic> map) {
-    final cm = reflectClass(T);
+  static dynamic invokeFromMap(Object Function() object, Map<String, dynamic> map) {
+    if (!isValidSchema(object)) throw InvalidSchemaException('Its given class does not extends from Schema!');
 
-    if (cm.declarations.containsKey(#fromMap)) {
-      return cm.invoke(#fromMap, [map]).reflectee;
+    final im = reflect(object());
+
+    if (im.type.declarations.containsKey(#fromMap)) {
+      final vms = im.type.superclass!.declarations.values.whereType<VariableMirror>().toList();
+
+      for (var vm in vms) {
+        if (vm.type.simpleName.name == 'DateTime' && map.containsKey(_snakeCaseNORMALIZER(vm.simpleName.name))) {
+          map.update(_snakeCaseNORMALIZER(vm.simpleName.name), (value) => (value as DateTime).millisecondsSinceEpoch);
+        }
+      }
+
+      return im.type.invoke(#fromMap, [map]).reflectee;
     } else {
       throw InvalidSchemaException('Its schema does not have a "fromMap" method!');
     }
   }
 
-  static dynamic invokeFromPostgreSQLMap(Type T, Map<String, dynamic> map) {
-    final cm = reflectClass(T);
+  static Map<String, dynamic> invokeToMap(Object Function() object) {
+    if (!isValidSchema(object)) throw InvalidSchemaException('Its given class does not extends from Schema!');
 
-    if (cm.declarations.containsKey(#fromPostgreSQLMap)) {
-      return cm.invoke(#fromPostgreSQLMap, [map]).reflectee;
-    } else {
-      throw InvalidSchemaException('Its schema does not have a "fromPostgreSQLMap" method!');
-    }
-  }
-
-  static Map<String, dynamic> invokeToMap(SQLetoSchema Function() schema) {
-    final im = reflect(schema());
+    final im = reflect(object());
 
     final cm = im.type;